Adding the code to build the IceBridge metadata hdf5 database.
The HDF5 database is used extensively throughout the IceBridge processing pipeline, as a way to get flightline metadata (lat, lon, elev, roll, pitch, yaw, etc...) from one file without having to traverse all the sub-directories and open hundreds-of-gigabytes of data just to pull metadata from it. It saves a lot of time to create this database once. The "IceBridgeGPR_Manager_v2.py" file uses this database extensively. This file traverses the directory structure and builds the database for you.

There also includes a member function for outputting a .shp file of icebridge flight lines for use in GIS visualizations. Code is as-is, and may need to be modified in the directory structure or file structure of the IceBridge files has changed since the code was written.